# backend/api/openapi/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Digital Store API",
    "description": "A digital product e-commerce system API providing endpoints for product browsing, shopping cart management, order processing, payment handling, and admin operations.",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "email": "support@digitalstore.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Type \"Bearer\" followed by a space and JWT token."
    }
  },
...
  "paths": {
...
    "/payments/online/session": {
      "post": {
        "tags": ["payments"],
        "summary": "Create online payment session",
        "description": "Create a Stripe checkout session for online payment. Response time < 500ms.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["order_id"],
              "properties": {
                "order_id": { "type": "integer", "example": 1 }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Payment session created with redirect URL" },
          "400": { "description": "Invalid order or already paid" }
        }
      }
    },
    "/payments/webhook/stripe": {
      "post": {
        "tags": ["payments"],
        "summary": "Stripe webhook",
        "description": "Handle Stripe payment webhooks. Verifies signature using HMAC-SHA256.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          { "in": "header", "name": "Stripe-Signature", "type": "string", "required": true }
        ],
        "responses": {
          "200": { "description": "Webhook processed" },
          "400": { "description": "Invalid signature or payload" }
        }
      }
    },
    "/payments/transfer/upload": {
      "post": {
        "tags": ["payments"],
        "summary": "Upload transfer proof",
        "description": "Upload a transfer payment proof image (JPEG, PNG, PDF). Max size: 10MB.",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          { "in": "formData", "name": "order_id", "type": "integer", "required": true },
          { "in": "formData", "name": "proof", "type": "file", "required": true }
        ],
        "responses": {
          "200": { "description": "Proof uploaded, order status updated to pending_transfer" },
          "400": { "description": "Invalid file or order" }
        }
      }
    },
...
  },
...
}